update UI for xiaomi service
