+++ front/src/routes/integration/all/xiaomi/sensorTh/index.js
@@ -2,20 +2,25 @@ import { Component } from 'preact';
 import { connect } from 'unistore/preact';
 import actions from './actions';
 import XiaomiSensorTemperaturePage from './XiaomiSensorTemperature';
+import XiaomiPage from '../XiaomiPage';
 
 @connect(
   'user,xiaomiSensorTemperature,houses,sensorTh',
   actions
 )
-class XiaomiIntegration extends Component {
+class XiaomiSensorThPage extends Component {
   componentWillMount() {
     this.props.getXiaomiSensorTemperature(100, 0);
     this.props.getHouses();
   }
 
   render(props, {}) {
-    return <XiaomiSensorTemperaturePage {...props} />;
+    return (
+      <XiaomiPage>
+        <XiaomiSensorTemperaturePage {...props} />
+      </XiaomiPage>
+    );
   }
 }
 
-export default XiaomiIntegration;
+export default XiaomiSensorThPage;
