+++ server/test/services/xiaomi/index.test.js
