+++ server/test/services/xiaomi/mocks.test.js
